requirements.txt
# requirements.txt
fastapi>=0.100.0
uvicorn==0.23.2
websockets==14.0
cryptography

streamlit>=1.25.0
python-dotenv>=1.0.0
upwork>=1.0.22  # Updated to use upwork package
httpx>=0.27.0 # Used for manual OAuth token exchange
pandas>=1.3.0 # For CSV export in frontend
requests-oauthlib==1.3.1
google-generativeai>=0.5.0